refactor: Allow non-vm local runtimes
