+++ internal/deploy/custom_slc_test.go
@@ -366,72 +366,6 @@ func TestUpsertInstalledCustomSLCReplacesAndSorts(t *testing.T) {
 	}
 }
 
-func TestLocalRunnerSlcArgsIncludesCustomMountAndPackage(t *testing.T) {
-	t.Parallel()
-
-	// Given
-	deployment := config.NewDeploymentDir(t.TempDir())
-	state := &config.ExasolPersonalState{
-		DeploymentVersion: "0.0.0",
-		InstalledSLCs: []config.InstalledSLC{
-			{Flavor: "python-3.12", Image: "img:py", Target: "/exa/slc/py"},
-		},
-		InstalledCustomSLCs: []config.InstalledCustomSLC{{
-			Alias:   "MYPY3",
-			Image:   "custom:mypy3-abc",
-			Target:  "/exa/slc/custom-mypy3",
-			Package: "custom-mypy3-abc.tar.gz",
-		}},
-	}
-	if err := config.WriteExasolPersonalState(state, deployment); err != nil {
-		t.Fatalf("failed to write state: %v", err)
-	}
-
-	// When
-	args, err := localRunnerSlcArgs(deployment)
-	// Then
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	want := []string{
-		"--slc", "img:py=/exa/slc/py",
-		"--slc", "custom:mypy3-abc=/exa/slc/custom-mypy3",
-		"--slc-package", "custom:mypy3-abc=custom-mypy3-abc.tar.gz",
-	}
-	if strings.Join(args, " ") != strings.Join(want, " ") {
-		t.Fatalf("expected official and custom args, got %v", args)
-	}
-}
-
-func TestLocalRunnerSlcArgsOmitsPackageForOfficialSLCs(t *testing.T) {
-	t.Parallel()
-
-	// Given
-	deployment := config.NewDeploymentDir(t.TempDir())
-	state := &config.ExasolPersonalState{
-		DeploymentVersion: "0.0.0",
-		InstalledSLCs: []config.InstalledSLC{
-			{Flavor: "python-3.12", Image: "img:py", Target: "/exa/slc/py"},
-		},
-	}
-	if err := config.WriteExasolPersonalState(state, deployment); err != nil {
-		t.Fatalf("failed to write state: %v", err)
-	}
-
-	// When
-	args, err := localRunnerSlcArgs(deployment)
-	// Then
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	for _, arg := range args {
-		if arg == "--slc-package" {
-			t.Fatalf("official SLCs must not be import-delivered: %v", args)
-		}
-	}
-}
-
 func TestConfirmOfficialAliasReuseBlocksInstalledOfficial(t *testing.T) {
 	t.Parallel()
