+++ internal/deploy/local_backend.go
@@ -53,7 +53,7 @@ var errUnsupportedLocalPlatform = errors.New(
 func newLocalBackend(
 	deployment config.DeploymentDir,
 	manifest *presets.InfrastructureManifest,
-	localRuntime localruntime.VMRuntime,
+	localRuntime localruntime.Runtime,
 ) *localBackend {
 	return &localBackend{deployment: deployment, manifest: manifest, runtime: localRuntime}
 }
@@ -61,7 +61,7 @@ func newLocalBackend(
 type localBackend struct {
 	deployment config.DeploymentDir
 	manifest   *presets.InfrastructureManifest
-	runtime    localruntime.VMRuntime
+	runtime    localruntime.Runtime
 }
 
 func (*localBackend) ValidateEnvironment() error {
